Restructuring and improvements.
